+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+## 1.1.0
+
+### 💡 Main Changes
+
+- [+] **@goldfishjs/react:** support React
+  - Add `useFetchInitData`, `useGlobalData`, `useGlobalFetchInitData`, `useGlobalReady`, `useLocalReady`, `usePlugins` and `usePluginsReady`.
+- [+] **@goldfishjs/utils:** add `isEqual`.
+
 ## 1.0.0
 
 ### 💡 Main Changes
@@ -5,7 +13,7 @@
 - [+] **@goldfishjs/react:** support React
   - Connect reactive system with React Function Component.
   - Import `useProps`, `useAutorun`, `useContextType`, `useState`, and `useWatch` from previous composition APIs.
-- [+] **@goldfishjs/pre-build:** Support `baseDir` and `tsconfigPath` configuration in `mini.project.json` file
+- [+] **@goldfishjs/pre-build:** support `baseDir` and `tsconfigPath` configuration in `mini.project.json` file
 
   ```json
   {